scanner script fixed
