+++ docs/code-samples/index.mdx
@@ -1,8 +1,3 @@
-
----
-title: Code Samples - XRPL.org
----
-
 import CodeSamples from "../../_components/CodeSamples.tsx";
 import { EmptyLayout } from "@redocly/developer-portal/ui";
