+++ package.json
@@ -12,9 +12,11 @@
   "dependencies": {
     "@redocly/developer-portal": "1.1.0-beta.101",
     "jquery": "^3.6.0",
+    "babel-plugin-preval": "^5.1.0",
     "fs": "^0.0.1-security",
     "jssoup": "^0.0.15",
-    "marked": "^4.0.18"
+    "marked": "^4.0.18",
+    "preval.macro": "^5.0.0"
   },
   "devDependencies": {
     "@types/jquery": "^3.5.14"
